requirements.txt
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.4.1
MarkupSafe==1.1.1
marshmallow==3.0.0
mock==3.0.5
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.1.0
python-dateutil==2.7.3
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.7
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.5.2
